enhancement: add new basic tests for ft_printf (beta)

+++ srcs/printf/ft_printf_d.c
@@ -6,7 +6,7 @@
 /*   By: jgigault <jgigault@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/01/06 11:34:07 by jgigault          #+#    #+#             */
-/*   Updated: 2015/01/06 17:08:28 by jgigault         ###   ########.fr       */
+/*   Updated: 2015/01/07 15:40:49 by jgigault         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ int				main(int argc, char **argv)
 {
 	int		ret;
 
-	ret = ft_printf(argv[1], atoi(argv[2]));
+	ret = ft_printf(argv[1], atol(argv[2]));
 	printf("|%d", ret);
 	return (0);
 }
