+++ srcs/printf/printf_d.c
@@ -6,7 +6,7 @@
 /*   By: jgigault <jgigault@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/01/06 11:34:07 by jgigault          #+#    #+#             */
-/*   Updated: 2015/01/06 17:07:57 by jgigault         ###   ########.fr       */
+/*   Updated: 2015/01/07 15:40:57 by jgigault         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ int				main(int argc, char **argv)
 {
 	int		ret;
 
-	ret = printf(argv[1], atoi(argv[2]));
+	ret = printf(argv[1], atol(argv[2]));
 	printf("|%d", ret);
 	return (0);
 }
